splitting vue renderer classes
